requirements_fixed.txt
altair==5.4.1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.2.2
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.6
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.14
django-cors-headers==4.3.1
django-extensions==3.2.3
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==2.2.0
drf-yasg==1.21.7
firebase-admin==6.5.0
Flask==3.0.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.2
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
narwhals==1.6.2
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.0
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
pydotplus==2.0.2
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smmap==5.0.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.5.0
streamlit==1.38.0
stripe==9.5.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.2
Werkzeug==3.0.3
whitenoise==6.6.0
